+++ php/wimb-options.php
@@ -44,21 +44,22 @@ function wimbblock_get_options_db() {
 function wimbblock_get_default_browsers() {
 	$defaults = array(
 		// https://developer.chrome.com/release-notes/
-		'Chrome'         => 139,
+		'Chrome'                   => 139,
 		// https://en.wikipedia.org/wiki/Microsoft_Edge#New_Edge_release_history
 		// https://www.cvedetails.com/version-list/26/32367/1/Microsoft-Edge.html?order=0
 		// Edge = Chrome
-		'Edge'           => 139,
+		'Edge'                     => 139,
 		// https://de.wikipedia.org/wiki/Versionsgeschichte_von_Mozilla_Firefox
 		// ESR  115.32.0  140.7.0
-		'Firefox'        => 140,
+		'Firefox'                  => 140,
 		// https://developer.apple.com/documentation/safari-release-notes
 		// https://theapplewiki.com/wiki/Safari
-		'Safari'         => 18,
+		'Safari'                   => 18,
 		// https://caniuse.com/usage-table
 		// SamsungBrowser/29.0 Chrome/136
 		// SamsungBrowser/28.0 Chrome/130
-		'SamsungBrowser' => 28,
+		// https://en.wikipedia.org/wiki/Samsung_Internet
+		'Samsung Internet Browser' => 28,
 	);
 	return $defaults;
 }